refactor(layout): extract column ordering
Move multi-column reading order and adjacent column merge helpers into a dedicated layout module so the layout entrypoint stays focused on constructing blocks.
